backend/requirements.txt
fastapi
uvicorn
python-dotenv
anthropic
langchain
langchain-anthropic
langchain-community
langchain-core
rank-bm25
pydantic
python-multipart
